code review: adding facade doc and moving analysis put/delete to analysis facade
